[FEAT/#11] 다이얼로그 구현
