+++ app/src/main/java/org/sopt/dosopttemplate/util/base/BindingDialogFragment.kt
@@ -4,10 +4,12 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.view.WindowManager
 import androidx.annotation.LayoutRes
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.DialogFragment
+import org.sopt.dosopttemplate.R
 
 abstract class BindingDialogFragment<T : ViewDataBinding>(
     @LayoutRes private val layoutResId: Int,
@@ -15,6 +17,17 @@ abstract class BindingDialogFragment<T : ViewDataBinding>(
     private var _binding: T? = null
     protected val binding: T
         get() = requireNotNull(_binding) { "binding object is not initialized" }
+
+    override fun onStart() {
+        super.onStart()
+        dialog?.window?.apply {
+            setLayout(
+                WindowManager.LayoutParams.MATCH_PARENT,
+                WindowManager.LayoutParams.WRAP_CONTENT,
+            )
+            setBackgroundDrawableResource(R.color.transparent)
+        }
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
